// client/stocklinx app/src/redux/branch/constant.ts
export enum branchConst {
  FETCH_BRANCHES_REQUEST = "FETCH_BRANCHES_REQUEST",
  FETCH_BRANCHES_SUCCESS = "FETCH_BRANCHES_SUCCESS",
  FETCH_BRANCHES_FAILURE = "FETCH_BRANCHES_FAILURE",
  FETCH_BRANCH_REQUEST = "FETCH_BRANCH_REQUEST",
  FETCH_BRANCH_SUCCESS = "FETCH_BRANCH_SUCCESS",
  FETCH_BRANCH_FAILURE = "FETCH_BRANCH_FAILURE",
  CREATE_BRANCH_REQUEST = "CREATE_BRANCH_REQUEST",
  CREATE_BRANCH_SUCCESS = "CREATE_BRANCH_SUCCESS",
  CREATE_BRANCH_FAILURE = "CREATE_BRANCH_FAILURE",
  UPDATE_BRANCH_REQUEST = "UPDATE_BRANCH_REQUEST",
  UPDATE_BRANCH_SUCCESS = "UPDATE_BRANCH_SUCCESS",
  UPDATE_BRANCH_FAILURE = "UPDATE_BRANCH_FAILURE",
  REMOVE_BRANCH_REQUEST = "REMOVE_BRANCH_REQUEST",
  REMOVE_BRANCH_SUCCESS = "REMOVE_BRANCH_SUCCESS",
  REMOVE_BRANCH_FAILURE = "REMOVE_BRANCH_FAILURE",
}